Added some more processing functions. Release version 0.2.0 in Vaadin Directory.

+++ src/main/java/org/vaadin/tltv/vprocjs/gwt/client/ui/test/ProcessingCodeImplTest.java
@@ -2,10 +2,9 @@ package org.vaadin.tltv.vprocjs.gwt.client.ui.test;
 
 import org.vaadin.tltv.vprocjs.gwt.client.ui.ProcessingCodeBase;
 
-
 public class ProcessingCodeImplTest extends ProcessingCodeBase {
 
-    float radius = 50.0f;
+    double radius = 50.0f;
     int X, Y;
     int nX, nY;
     int delay = 16;
